ClojureBindings: implement all accessor methods
Previously, they threw UnsupportedOperationException due to the
fact that it is tricky to access the list of available user keys.

But it _can_ be done, so let's do it!

This makes Clojure work much better in the SciJava REPL, which needs
the ability to iterate through the variables of the engine bindings.