ClojureBindings: avoid warnings in map population
Calling the Var.intern method dumps a warning to stderr
when an already-existing variable is requested.

I did not study the code extremely carefully, but it seems like
internally, Clojure discovers an existing variable, decides to overwrite
it with a different variable object for some reason, and then warns that
it has done so. But this phenomenon apparently occurs only once.

In any case, the workaround is simply to avoid calling Var.intern,
in favor of Namespace.getMappings().valAt(Symbol) directly.